requirements.txt
fastapi==0.111.1
pydantic==2.11.0
python-multipart==0.0.9
uvicorn==0.30.3
gunicorn==22.0.0
requests==2.32.3
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
pdf-annotate==0.12.0
scipy==1.14.0
opencv-python==4.10.0.84
Shapely==2.0.5
transformers==4.40.2
huggingface_hub==0.23.5
pdf2image==1.17.0
lightgbm==4.5.0
setuptools==75.4.0
roman==4.2
hydra-core==1.3.2
pypandoc==1.13
rapid-table==2.0.3
rapidocr==3.2.0
pix2tex==0.1.4
latex2mathml==3.78.0
PyMuPDF==1.25.5
ollama==0.6.0
openai==1.57.4
cachetools==6.2.1
git+https://github.com/huridocs/pdf-features.git@2025.10.16.3